refactor: consolidate and enhance unit tests for subnet and pool status mappers
- Removed the obsolete mapper_test.go file and replaced it with new tests in subnet_mapper_test.go and poolstatus_mapper_test.go.
- Added comprehensive test cases for the mapSubnetToParentPool and mapAllocToPool functions, ensuring proper handling of PoolRef and namespace validations.
- Introduced additional tests for duplicate request suppression in workqueue_dedupe_test.go, verifying the behavior of the workqueue under various conditions.
